swich to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[]{1,12,343,676,878};
+        long[] sales = new long[]{1,12,343,676,878};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.print(max);
     }
 }
